Added TypeScript integration -changed files extentions;
