+++ rn-cli.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  getTransformModulePath() {
+    return require.resolve('react-native-typescript-transformer')
+  },
+  getSourceExts() {
+    return ['ts', 'tsx']
+  },
+}
