+++ src/components/NoInternetIndicator.tsx
