+++ src/components/StyledInput.tsx
