+++ src/hooks/typedReduxHooks.ts
@@ -1,0 +1,10 @@
+import {
+  TypedUseSelectorHook,
+  useDispatch as useReduxDispatch,
+  useSelector as useReduxSelector,
+} from 'react-redux'
+
+import type { RootState, AppDispatch } from '../redux/store'
+
+export const useDispatch = () => useReduxDispatch<AppDispatch>()
+export const useSelector: TypedUseSelectorHook<RootState> = useReduxSelector
