+++ src/redux/store.ts
@@ -10,4 +10,7 @@ const store = configureStore({
   },
 })
 
+export type RootState = ReturnType<typeof store.getState>
+export type AppDispatch = typeof store.dispatch
+
 export default store
