+++ src/screens/Login/index.tsx
@@ -1,23 +1,22 @@
 import React from 'react'
-import { useDispatch } from 'react-redux'
 import { useForm } from 'react-hook-form'
 import { yupResolver } from '@hookform/resolvers/yup'
-import { ScrollView, View, Text, Pressable } from 'react-native'
+import { ScrollView, View, Text, Pressable, Alert } from 'react-native'
 
-import { signUpUser } from '../../services/user'
+import { useDispatch } from '../../hooks'
+import { loginUser } from '../../services/user'
 import Routes from '../../routes'
-import { signUpValidationSchema } from '../../utils/validators'
+import { loginValidationSchema } from '../../utils/validators'
 import theme from '../../theme'
 import Layout from '../../components/Layout'
 import StyledFormInput from '../../components/StyledFormInput'
 import cs from './styles'
 
-const SignUp = ({ navigation, route }) => {
+const Login = ({ navigation }) => {
   const dispatch = useDispatch()
 
   const formDefaultValues = {
     username: '',
-    email: '',
     password: '',
   }
 
@@ -27,11 +26,19 @@ const SignUp = ({ navigation, route }) => {
     formState: { errors },
   } = useForm({
     defaultValues: formDefaultValues,
-    resolver: yupResolver(signUpValidationSchema),
+    resolver: yupResolver(loginValidationSchema),
   })
 
+  const onSignupPress = () => {
+    navigation.navigate(Routes.SignUp)
+  }
+
+  const onForgotPasswordPress = () => {
+    Alert.alert('Forgot password', 'Not implemeted yet :)')
+  }
+
   const onSubmit = async (data) => {
-    signUpUser(data)(dispatch)
+    loginUser(data)(dispatch)
   }
 
   return (
@@ -43,7 +50,7 @@ const SignUp = ({ navigation, route }) => {
       >
         <View style={theme.helpers.flex1} />
         <View style={theme.helpers.flex1JustifyContentCenter}>
-          <Text style={[cs.titleBlock, theme.textStyles.title1]}>Sign Up</Text>
+          <Text style={[cs.titleBlock, theme.textStyles.title1]}>Login</Text>
           <View style={theme.helpers.ph20}>
             <StyledFormInput
               style={theme.helpers.mb16}
@@ -51,26 +58,18 @@ const SignUp = ({ navigation, route }) => {
               control={control}
               errors={errors}
             />
-            <StyledFormInput
-              style={theme.helpers.mb16}
-              name="email"
-              control={control}
-              errors={errors}
-            />
             <StyledFormInput name="password" control={control} errors={errors} />
             <Pressable style={cs.submitButton} onPress={handleSubmit(onSubmit)}>
-              <Text style={[theme.textStyles.title3, cs.whiteColor]}>Sign Up</Text>
+              <Text style={[theme.textStyles.title3, cs.whiteColor]}>Log In</Text>
             </Pressable>
           </View>
         </View>
         <View style={theme.helpers.alignItemsCenter}>
-          <Pressable style={theme.helpers.pb20} onPress={() => navigation.navigate(Routes.Login)}>
-            <Text style={theme.textStyles.subHeadline}>
-              Already have an account ?{' '}
-              <Text style={[theme.textStyles.subHeadlineBold, cs.pressableTextColor]}>
-                Sign In.
-              </Text>
-            </Text>
+          <Pressable style={theme.helpers.pb20} onPress={onSignupPress}>
+            <Text style={[theme.textStyles.subHeadlineBold, cs.pressableTextColor]}>Sign Up</Text>
+          </Pressable>
+          <Pressable style={theme.helpers.pb20} onPress={onForgotPasswordPress}>
+            <Text style={theme.textStyles.subHeadline}>Forgot password ?</Text>
           </Pressable>
         </View>
       </ScrollView>
@@ -78,4 +77,4 @@ const SignUp = ({ navigation, route }) => {
   )
 }
 
-export default SignUp
+export default Login
