+++ src/screens/SplashScreen.tsx
