+++ src/types/user.ts
@@ -1,0 +1,1 @@
+export type CurrentUser = any
